Moved random username generation to the backend
