+++ services/agora/src/pages/onboarding/step4-username/index.vue
@@ -14,7 +14,8 @@
           </div>
 
           <q-input v-model="userName" label="Username" outlined :maxlength="MAX_LENGTH_USERNAME"
-            :error="!isValidUsername" :error-message="userNameInvalidMessage">
+            :error="!isValidUsername" :error-message="userNameInvalidMessage"
+            @update:model-value="nameContainsValidCharacters">
             <template #append>
               <ZKButton v-if="isValidUsername" icon="mdi-check" text-color="red" />
               <ZKButton icon="mdi-dice-6" @click="refreshName()" />
@@ -37,10 +38,8 @@
 import StepperLayout from "src/components/onboarding/StepperLayout.vue";
 import InfoHeader from "src/components/onboarding/InfoHeader.vue";
 import { useRouter } from "vue-router";
-import { ref, watch } from "vue";
+import { onMounted, ref } from "vue";
 import ZKButton from "src/components/ui-library/ZKButton.vue";
-import { useBackendOnboardingApi } from "src/utils/api/onboarding";
-import { generateRandomUsername } from "src/shared/services/account";
 import { MAX_LENGTH_USERNAME } from "src/shared/shared";
 import { zodUsername } from "src/shared/types/zod";
 import { ZodError } from "zod";
@@ -48,28 +47,33 @@ import { useBackendAccountApi } from "src/utils/api/account";
 
 const router = useRouter();
 
+// const { showNotifyMessage } = useNotify();
+
 const userNameInvalidMessage = ref("");
 const isValidUsername = ref(true);
 
-const { isUsernameInUse } = useBackendOnboardingApi();
-const { submitUsernameChange } = useBackendAccountApi();
+const {
+  submitUsernameChange,
+  isUsernameInUse,
+  generateUnusedRandomUsername
+} = useBackendAccountApi();
 
 const validationMessage = ref("");
 
-const userName = ref(generateRandomUsername());
+const userName = ref("");
 
-watch(userName, () => {
-  nameContainsValidCharacters(userName.value);
+onMounted(async () => {
+  userName.value = await generateUnusedRandomUsername();
 })
 
-async function nameContainsValidCharacters(value: string): Promise<boolean> {
+async function nameContainsValidCharacters(): Promise<boolean> {
   try {
-    zodUsername.parse(value);
+    zodUsername.parse(userName.value);
 
-    const isInUse = await isUsernameInUse(value);
+    const isInUse = await isUsernameInUse(userName.value);
     if (isInUse) {
       isValidUsername.value = false;
-      userNameInvalidMessage.value = "The entered username is currently in use"
+      userNameInvalidMessage.value = "The entered username is already in use"
       return false;
     } else {
       isValidUsername.value = true;
@@ -84,8 +88,8 @@ async function nameContainsValidCharacters(value: string): Promise<boolean> {
   }
 }
 
-function refreshName() {
-  userName.value = generateRandomUsername();
+async function refreshName() {
+  userName.value = await generateUnusedRandomUsername();
 }
 
 async function goToNextRoute() {
